s5w1 (feat): update delete all data endpoint

+++ src/common/tests/utils/delete-all-data.ts
@@ -1,14 +1,17 @@
 import { Connection } from 'mongoose'
 import { DataSource } from 'typeorm'
 
-export const deleteAllData = async (databaseConnection: Connection, dataSource: DataSource) => {
-  await databaseConnection.collection('users').deleteMany({})
-  await databaseConnection.collection('blogs').deleteMany({})
-  await databaseConnection.collection('posts').deleteMany({})
-  await databaseConnection.collection('comments').deleteMany({})
-  await databaseConnection.collection('likes').deleteMany({})
+export const deleteAllData = async (databaseConnection: Connection | null, dataSource: DataSource | null) => {
+  if (databaseConnection) {
+    await databaseConnection.collection('users').deleteMany({})
+    await databaseConnection.collection('blogs').deleteMany({})
+    await databaseConnection.collection('posts').deleteMany({})
+    await databaseConnection.collection('comments').deleteMany({})
+    await databaseConnection.collection('likes').deleteMany({})
+  }
 
-  await dataSource.query(`
+  if (dataSource) {
+    await dataSource.query(`
     DO $$
       BEGIN
         IF EXISTS (SELECT 1 FROM information_schema.tables WHERE table_schema = 'public' AND table_name = 'users') THEN
@@ -16,7 +19,7 @@ export const deleteAllData = async (databaseConnection: Connection, dataSource: 
         END IF;
     END $$;
   `)
-  await dataSource.query(`
+    await dataSource.query(`
     DO $$
       BEGIN
         IF EXISTS (SELECT 1 FROM information_schema.tables WHERE table_schema = 'public' AND table_name = 'devices') 
@@ -25,4 +28,5 @@ export const deleteAllData = async (databaseConnection: Connection, dataSource: 
         END IF;
     END $$;
   `)
+  }
 }
